fix pipeline

+++ sneakPeekCrawler/pipelines.py
@@ -26,8 +26,10 @@ class SneakpeekcrawlerPipeline(object):
         self.db = self.client.collection(self.collection_name)
 
     def process_item(self, item, spider):
-        result = self.db.add(dict(item))
-        if utils.isDevelopment():
-            self.db.document(result[1].id).delete()
+        records = list(self.db.where(u'image_urls', u'==', dict(item)[
+            "image_urls"]).get())
+        if not bool(records):
+            result = self.db.add(dict(item))
+            if utils.isDevelopment():
+                self.db.document(result[1].id).delete()
         return item
-
